requirements.txt
accelerate==0.27.2
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.25.7
anyio==4.2.0
appdirs==1.4.4
asgiref==3.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blessed==1.20.0
build==1.2.2.post1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
click-default-group==1.2.4
cloudpickle==3.0.0
cmake==3.29.6
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
comm @ file:///croot/comm_1709322850197/work
compressed-tensors==0.8.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
datasets==2.17.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepspeed==0.13.2
Deprecated==1.2.15
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
durationpy==0.9
einops==0.7.0
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss-cpu==1.9.0.post1
fastapi==0.115.0
ffmpy==0.3.2
filelock==3.13.1
fire==0.6.0
flash-attn==2.5.8
flatbuffers==24.3.25
fonttools==4.49.0
Fraction==2.2.0
frozenlist==1.4.1
fschat==0.2.36
fsspec==2023.10.0
func-timeout==4.3.5
fuzzywuzzy==0.18.0
gdown==5.2.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.62.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gpustat==1.1.1
gradio==4.21.0
gradio_client==0.12.0
greenlet==3.1.1
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.3
h11==0.14.0
hjson==3.1.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.6
import==0.0.1
importlib_metadata==8.4.0
importlib_resources==6.1.2
iniconfig==2.0.0
interegular==0.3.3
ipykernel @ file:///croot/ipykernel_1728665589812/work
ipython @ file:///croot/ipython_1726064227985/work
isort==5.13.2
jedi @ file:///croot/jedi_1721058342488/work
jieba==0.42.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.3.2
jq==1.8.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.8
langchain-chroma==0.1.4
langchain-community==0.3.8
langchain-core==0.3.21
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langsmith==0.1.145
lark==1.1.9
lark-oapi==1.2.1
lazy_loader==0.4
Levenshtein==0.25.1
librosa==0.10.2.post1
linkify-it-py==2.0.3
llvmlite==0.43.0
lm-format-enforcer==0.10.9
loralib==0.1.2
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.8.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.13.4
mistral_common==1.5.0
mmh3==5.0.1
modelscope==1.18.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.2.1
nh3==0.2.15
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.47.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
-e git+ssh://git@github.com/OpenLLMAI/OpenRLHF.git@df92b437a34ec832648a275949f493574dcded68#egg=openrlhf
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
optimum==1.16.2
orjson==3.9.15
outlines==0.0.43
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partial-json-parser==0.2.1.1.post4
peft==0.10.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy==1.5.0
pooch==1.8.2
posthog==3.7.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
proto-plus==1.24.0
protobuf==5.28.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
psycopg2-binary==2.9.9
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
py-spy==0.3.14
pyairports==2.1.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments @ file:///croot/pygments_1684279966437/work
PyMySQL==1.1.1
pyparsing==3.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.1
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
rapidfuzz==3.9.4
ray==2.35.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.0
rouge-chinese==1.0.3
rpds-py==0.18.0
rsa==4.9
ruff==0.2.2
safetensors==0.4.2
scikit-learn==1.5.1
scipy==1.12.0
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==1.40.5
setproctitle==1.3.3
shapely==2.0.5
shellingham==1.5.4
shortuuid==1.0.11
shtab==1.7.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==7.0.1
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0
SQLAlchemy==2.0.35
sqlglot==25.5.1
sqlite-fts4==1.0.3
sqlite-utils==3.37
sqlparse==0.5.1
sse-starlette==2.1.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.38.5
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboardX==2.6.2.2
termcolor==2.4.0
textual==0.78.0
thefuzz==0.22.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
tokenizers==0.20.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.1
torchvision==0.20.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.2
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.46.3
transformers-stream-generator==0.0.4
triton==3.1.0
trl==0.8.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tyro==0.7.3
tzdata==2023.4
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
virtualenv==20.25.1
vllm==0.6.4.post1
vllm-flash-attn==2.6.1
wandb==0.16.3
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
wrapt==1.16.0
xformers==0.0.28.post3
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.1
